+++ packages/liveblocks-node/src/__tests__/client.test.ts
@@ -7,6 +7,7 @@ import type {
   RoomSubscriptionSettings,
   StorageNode,
   ThreadData,
+  ThreadVisibility,
 } from "@liveblocks/core";
 import { createNotificationSettings, LiveList } from "@liveblocks/core";
 import { http, HttpResponse } from "msw";
@@ -931,6 +932,7 @@ describe("client", () => {
 
   describe("create thread", () => {
     test("should return the created thread when createThread receives a successful response", async () => {
+      const visibility: ThreadVisibility = "private";
       const threadData = {
         comment: {
           userId: "user1",
@@ -943,7 +945,7 @@ describe("client", () => {
         metadata: {
           color: "blue",
         },
-        visibility: "private",
+        visibility,
       };
 
       server.use(